+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @abloatai/cli
 
+## 0.59.1
+
 ## 0.59.0
 
 ### Minor Changes